Extracteed DisplayableState into a separate file.
